relocate files to re-align with current best-practices
