+++ src/sections/shared/layout/SeparationLine/SeparationLine.module.scss
